Conflicts:
	backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/persistence/repository/RoleRepository.java
	backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/rest/controller/RoleController.java
	backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/rest/handler/RoleHandler.java
	backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/rest/handler/impl/RoleHandlerImpl.java
	backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/service/RoleService.java
	backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/service/impl/RoleServiceImpl.java

+++ backend/judicial-warrant/src/main/java/com/informatique/gov/judicialwarrant/support/modelmpper/CandidateAttachmentMapper.java
@@ -14,7 +14,7 @@ import lombok.AllArgsConstructor;
 @Component
 public class CandidateAttachmentMapper extends AbstractModelMapper<CandidateAttachment, CandidateAttachmentDto, Long> {
 
-	private ModelMapper<AttachmentType, AttachmentTypeDto, Byte> attachmentTypeMapper;
+	private ModelMapper<AttachmentType, AttachmentTypeDto, Long> attachmentTypeMapper;
 	private ModelMapper<Candidate, CandidateDto, Long> candidateMapper;
 
 	/**
@@ -32,9 +32,7 @@ public class CandidateAttachmentMapper extends AbstractModelMapper<CandidateAtta
 			dto.setUcmDocumentId(entity.getUcmDocumentId());
 			dto.setAttachmentType(attachmentTypeMapper.toDto(entity.getAttachmentType()));
 			dto.setCandidate(candidateMapper.toDto(entity.getCandidate()));
-
 		}
-
 		return dto;
 	}
